update and delete in patient and edit doctor

+++ graduation project/app/app/src/main/java/app/app/Repositories/DoctorRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DoctorRepo extends CrudRepository<Doctor, Long> {
+public interface DoctorRepo extends CrudRepository<Doctor, Integer> {
     Doctor findByEmail(String email);
 }
